refactor(blocks-engine): move the op vocabulary to the engine, builder re-exports
WORK IN PROGRESS, held on this branch pending a decision. The move is complete
and proven pure, but it cannot pass the quality gate for a structural reason
recorded alongside it.

The vocabulary lives in its own engine module rather than in operations.ts. The
plan named that file, but format.ts re-exports it and the vocabulary needs the
registry and the validators, which reach picomatch and css-tree. That entry
point exists so a generator or a schema publisher does not load the validator,
the migrations and the CSS parser, and its own docblock predicts exactly this
regression. Its boundary test caught it.

The relocation is byte-identical after the import block, and the 202 existing
ops tests pass unchanged through the builder re-export.
